+++ .gitignore
@@ -1,3 +1,5 @@
-__pycache__/
+*.egg-info/
 .venv/
+__pycache__/
+build/
 venv/
