+++ setup.cfg
@@ -8,6 +8,14 @@ maintainer = SigOpt
 maintainer_email = support@sigopt.com
 long_description = Tools for maintaining SigOpt repositories
 
+[options]
+python_requires = >=3.7
+packages = find_namespace:
+
+[options.packages.find]
+exclude=
+  test*
+
 [flake8]
 ignore=E129,E127,E203,E302,E131,E111,E114,E121,E501,E126,E123,E305,E402,I101,I100,N806,F403,E241,E731,F999,F401,F405,W503,E741,W504
