+++ sigopt_tools/check_copyright_and_license_disclaimers.py
