+++ sigopt_tools/shell_lint.py
